attached_assets/config.py
# Python file to load environment variables
import os
from dotenv import load_dotenv

def load_env_variables():
    """Load environment variables from .env file"""
    load_dotenv()
    
    # Get environment variables
    client_id = os.getenv("SPOTIFY_CLIENT_ID")
    client_secret = os.getenv("SPOTIFY_CLIENT_SECRET")
    redirect_uri = os.getenv("REDIRECT_URI")
    
    return client_id, client_secret, redirect_uri
