attached_assets/requirements.txt
streamlit==1.32.2
spotipy==2.23.0
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.4
seaborn==0.13.1
pillow==10.2.0
python-dotenv==1.0.0